Ensure removal of FDs from epoll after they fire
Since we store arbitrary types behind a raw pointer in epoll, we don't
want to alias the pointer. When an event fires, we reconstruct the type
into a Box, but the Box invariants don't hold since there is an alias
still registered with epoll. We must remove the event altogether so that
there is only one instance of the pointer owned by the Box.

Also, start listening in the main thread of the multithreaded example
test to prevent timing issues.