Add reregister/deregister commands to Registrar
Allow actively changing the event type and deleting registrations from
the kernel poller when an event for a file descriptor has not yet fired.
This requires some sort of coordination between the register and poller
to make this safe. In order to avoid mutexes, this is done via Atomics.

This commit only makes the change for kqueue and does not add any
testing. As such it is WIP and not ready for merge.