Disallow arbitrary types as userdata
Amy no-longer allows arbitrary types to be registered as userdata. This
design decision came out of the scenario described in #3.

This vastly simplifies the code and removes the burden of performing
allocations for each registration and notification. Now the only
allocations done are at the time the Poller is created, in order to
provide buffer space for events from the kernel. This makes Amy much
faster and more predictable.

Removing userdata also removes the need to maintain Registration and
Socket types, because the worker thread maintains ownership over the
original socket tied to a unique ID returned on registration. This
allows a-priori knowledge of socket types and no need for that
information to be returned from the poller.

Furthermore, there is no longer a need for the Poller and Registrar to
be generic.

Finally, there was significant removal of unsafe code, with the only
unsafe code being utilized to resize a vec after using it across an FFI
boundary with a poller.