Add pollable channel receiver support
Async and sync channel receivers can now be polled by epoll/kqueue.
Since syscalls are very expensive and are used for user level
notifications in both kqueue and epoll, we reduce the need for them.
This is accomplished via an atomic `pending` counter that gets
incremented on every send and decremented on every receive. When the
counter is 0 during either a send or receive, a syscall will be made to
trigger or clear the event in the kernel poller respectively.

Note that the above optimization only helps when there are multiple
sends between polls on the same channel. It is therefore more efficient
to use a single channel between the poller and sending thread then
multiple channels.