Ensure user events get retriggered after clear
Channels were not properly waking up the poller on kqueue based systems
when a send was made after a failing try_recv() call. It turns out that
on kqueue based systems EV_DISABLE (which is set on clearing of events)
is maintained even when an event is 're-added' with EV_ADD later, such
as during a user event trigger. The fix was to re-add the event with
EV_ADD | EV_ENABLE.

A regression test was added for the fix as well.

Bump version to 0.5.6