Make channel eventfd usage safer on Linux
Previously, the channel sender and receiver contained a copy of the same RawFd
If the channel was closed this RawFd would leak. Fixing that leak by calling
close() on the fd could actually result in erroneous behavior. Eventfd files
signal user events via reading and writing. If the Fd was closed and re-used by
the kernel, writing to it with the sender could potentially overwrite an unrelated
file. We now do the right thing and dup the fd so that the race condition doesn't exist.
When all senders and receivers are dropped, all duplicate fds are closed and the
underlying file is cleaned up.