All timers are now managed by the poller on linux
A new timer channel was created so that timer_ids could be created by the user
at the registrars but sent directly to the poller thread for actual creation of
timerfd file descriptors and registration with epoll. These timers are also tracked
so that when a notification occurs they can be re-armed in the poller rather than
requiring the client of amy to do this. This simplifies the API and helps prevent
mistakes where interval timers don't fire because the user did not re-arm them.

While this change makes the API more ergonomic, it is slightly more expensive in that
it requires a channel send and extra tracking and checking at the poller. This seems to
be a reasonable tradeoff however, as registering timers should be very infrequent. Timers
should only be coarse grained and used to drive timer wheels for finer granularity and
greater scale. The tracking is unfortunate, but cost may be mitigated in the future by
storing the type of notification in the high bits of the IDs, so that extra hashmap
lookups are not required on each poll result.

All of the above is very useful, but the change was primarily motivated in order to
provide a method for creating timers that do not require timerfd usage. This new channel
can be used to instruct the poller to adjust it's timeouts to the minimum timer expiry so
that this feature can be implemented. The reason for implementation is that some linux
emulation layers and old versions of linux do not support timerfd.

Additionally, there were some other relatively minor changes:
 * Timers are now just usize IDs like other notifications from a user POV
 * Registrars are no longer clone, as they must clone an amy channel, which itself
   requires a UserEvent clone which requires a `libc::dup` call which can fail. There
   is a new `try_clone(&self)` method as a replacement.
 * The timer is now only internally used by epoll and it's structure is different.
 * All errorable return types from KernelPoller, channels, timers etc.. are `io::Result`
   and no longer `nix::Result`
 * Tests were fixed to reflect the new API