Don't deregister user event on drop for kqueue
For kqueue based systems, there is no file descriptor representing each
user event. There is just a unique id that gets copied, rather than
`dup`ed each time the user_event is copied. For channels, this means
that if a sender goes away, if its user event was dropped it would
deregister the event with kqueue. Since this event id was shared between
other senders and the receiver, the channel becomes unpollable.
Instead, we only deregister the user event when the channel receiver is
dropped. Since user events are not in the public api and only currently
used for channels, this works well.